requirements.txt
selenium
deep-translator
python-dotenv